test: move ITs to engine tests
ITs are very slow, these tests are fine as engine tests. As long as we have a single scenario as IT that is sufficient.
I have chosen to keep the deployment authorization IT here. It is also copied because I will modify it in a later commit
.
