+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/DecisionEvaluationEvaluateAuthorizationIT.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.it.authorization;
+package io.camunda.zeebe.engine.processing.authorization.permissions;
 
 import static io.camunda.zeebe.it.util.AuthorizationsUtil.createClient;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -18,7 +18,6 @@ import io.camunda.client.protocol.rest.PermissionTypeEnum;
 import io.camunda.client.protocol.rest.ResourceTypeEnum;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
-import io.camunda.zeebe.model.bpmn.Bpmn;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
@@ -26,6 +25,7 @@ import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
 import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -36,12 +36,13 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class ProcessInstanceCreateAuthorizationIT {
+public class DecisionEvaluationEvaluateAuthorizationIT {
+
   @Container
   private static final ElasticsearchContainer CONTAINER =
       TestSearchContainers.createDefeaultElasticsearchContainer();
 
-  private static final String PROCESS_ID = "processId";
+  private static final String DECISION_ID = "jedi_or_sith";
   private static AuthorizationsUtil authUtil;
   @AutoCloseResource private static CamundaClient defaultUserClient;
 
@@ -64,41 +65,28 @@ public class ProcessInstanceCreateAuthorizationIT {
     authUtil.awaitUserExistsInElasticsearch(defaultUsername);
     defaultUserClient
         .newDeployResourceCommand()
-        .addProcessModel(
-            Bpmn.createExecutableProcess(PROCESS_ID).startEvent().endEvent().done(), "process.xml")
+        .addResourceFromClasspath("dmn/drg-force-user.dmn")
         .send()
         .join();
   }
 
   @Test
-  void shouldBeAuthorizedToCreateInstanceWithDefaultUser() {
-    // when then
-    final var processInstanceEvent =
+  void shouldBeAuthorizedToEvaluateDecisionWithDefaultUser() {
+    // when
+    final var response =
         defaultUserClient
-            .newCreateInstanceCommand()
-            .bpmnProcessId(PROCESS_ID)
-            .latestVersion()
+            .newEvaluateDecisionCommand()
+            .decisionId(DECISION_ID)
+            .variables(Map.of("lightsaberColor", "red"))
             .send()
             .join();
-    assertThat(processInstanceEvent.getBpmnProcessId()).isEqualTo(PROCESS_ID);
+
+    // then
+    assertThat(response.getDecisionOutput()).isEqualTo("\"Sith\"");
   }
 
   @Test
-  void shouldBeAuthorizedToCreateInstanceWithResultWithDefaultUser() {
-    // when then
-    final var processInstanceResult =
-        defaultUserClient
-            .newCreateInstanceCommand()
-            .bpmnProcessId(PROCESS_ID)
-            .latestVersion()
-            .withResult()
-            .send()
-            .join();
-    assertThat(processInstanceResult.getBpmnProcessId()).isEqualTo(PROCESS_ID);
-  }
-
-  @Test
-  void shouldBeAuthorizedToCreateInstanceWithUser() {
+  void shouldBeAuthorizedToEvaluateDecisionWithUser() {
     // given
     final var username = UUID.randomUUID().toString();
     final var password = "password";
@@ -106,69 +94,27 @@ public class ProcessInstanceCreateAuthorizationIT {
         username,
         password,
         new Permissions(
-            ResourceTypeEnum.PROCESS_DEFINITION,
-            PermissionTypeEnum.CREATE_PROCESS_INSTANCE,
-            List.of(PROCESS_ID)));
+            ResourceTypeEnum.DECISION_DEFINITION,
+            PermissionTypeEnum.CREATE_DECISION_INSTANCE,
+            List.of(DECISION_ID)));
 
     try (final var client = authUtil.createClient(username, password)) {
-      // when then
-      final var processInstanceEvent =
-          client.newCreateInstanceCommand().bpmnProcessId(PROCESS_ID).latestVersion().send().join();
-      assertThat(processInstanceEvent.getBpmnProcessId()).isEqualTo(PROCESS_ID);
-    }
-  }
-
-  @Test
-  void shouldBeAuthorizedToCreateInstanceWithResultWithUser() {
-    // given
-    final var username = UUID.randomUUID().toString();
-    final var password = "password";
-    authUtil.createUserWithPermissions(
-        username,
-        password,
-        new Permissions(
-            ResourceTypeEnum.PROCESS_DEFINITION,
-            PermissionTypeEnum.CREATE_PROCESS_INSTANCE,
-            List.of(PROCESS_ID)));
-
-    try (final var client = authUtil.createClient(username, password)) {
-      // when then
-      final var processInstanceResult =
+      // when
+      final var response =
           client
-              .newCreateInstanceCommand()
-              .bpmnProcessId(PROCESS_ID)
-              .latestVersion()
-              .withResult()
+              .newEvaluateDecisionCommand()
+              .decisionId(DECISION_ID)
+              .variables(Map.of("lightsaberColor", "red"))
               .send()
               .join();
-      assertThat(processInstanceResult.getBpmnProcessId()).isEqualTo(PROCESS_ID);
-    }
-  }
-
-  @Test
-  void shouldBeUnauthorizedToCreateInstanceIfNoPermissions() {
-    // given
-    final var username = UUID.randomUUID().toString();
-    final var password = "password";
-    authUtil.createUser(username, password);
-
-    try (final var client = authUtil.createClient(username, password)) {
-      // when
-      final var createFuture =
-          client.newCreateInstanceCommand().bpmnProcessId(PROCESS_ID).latestVersion().send();
 
       // then
-      assertThatThrownBy(createFuture::join)
-          .isInstanceOf(ProblemException.class)
-          .hasMessageContaining("title: FORBIDDEN")
-          .hasMessageContaining("status: 403")
-          .hasMessageContaining(
-              "Insufficient permissions to perform operation 'CREATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION'");
+      assertThat(response.getDecisionOutput()).isEqualTo("\"Sith\"");
     }
   }
 
   @Test
-  void shouldBeUnauthorizedToCreateInstanceWithResultIfNoPermissions() {
+  void shouldBeUnauthorizedToEvaluateDecisionIfNoPermissions() {
     // given
     final var username = UUID.randomUUID().toString();
     final var password = "password";
@@ -176,22 +122,21 @@ public class ProcessInstanceCreateAuthorizationIT {
 
     try (final var client = authUtil.createClient(username, password)) {
       // when
-      final var createFuture =
+      final var response =
           client
-              .newCreateInstanceCommand()
-              .bpmnProcessId(PROCESS_ID)
-              .latestVersion()
-              .withResult()
+              .newEvaluateDecisionCommand()
+              .decisionId(DECISION_ID)
+              .variables(Map.of("lightsaberColor", "red"))
               .send();
 
       // then
-      assertThatThrownBy(createFuture::join)
+      assertThatThrownBy(response::join)
           .isInstanceOf(ProblemException.class)
           .hasMessageContaining("title: FORBIDDEN")
           .hasMessageContaining("status: 403")
           .hasMessageContaining(
-              "Insufficient permissions to perform operation 'CREATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
-              PROCESS_ID);
+              "Insufficient permissions to perform operation 'CREATE_DECISION_INSTANCE' on resource 'DECISION_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              DECISION_ID);
     }
   }
 }
