+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/JobCompleteAuthorizationIT.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.it.authorization;
+package io.camunda.zeebe.engine.processing.authorization.permissions;
 
 import static io.camunda.zeebe.it.util.AuthorizationsUtil.createClient;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -19,10 +19,7 @@ import io.camunda.client.protocol.rest.ResourceTypeEnum;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
 import io.camunda.zeebe.model.bpmn.Bpmn;
-import io.camunda.zeebe.protocol.record.RejectionType;
-import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
-import io.camunda.zeebe.protocol.record.value.BpmnElementType;
-import io.camunda.zeebe.protocol.record.value.BpmnEventType;
+import io.camunda.zeebe.protocol.record.intent.JobIntent;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
@@ -41,8 +38,9 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class SignalBroadcastAuthorizationIT {
-  public static final String SIGNAL_NAME = "signal";
+public class JobCompleteAuthorizationIT {
+
+  public static final String JOB_TYPE = "jobType";
 
   @Container
   private static final ElasticsearchContainer CONTAINER =
@@ -74,12 +72,8 @@ public class SignalBroadcastAuthorizationIT {
         .addProcessModel(
             Bpmn.createExecutableProcess(PROCESS_ID)
                 .startEvent()
-                .intermediateCatchEvent()
-                .signal(s -> s.name(SIGNAL_NAME))
+                .serviceTask("serviceTask", t -> t.zeebeJobType(JOB_TYPE))
                 .endEvent()
-                .moveToProcess(PROCESS_ID)
-                .startEvent()
-                .signal(s -> s.name(SIGNAL_NAME))
                 .done(),
             "process.xml")
         .send()
@@ -87,22 +81,32 @@ public class SignalBroadcastAuthorizationIT {
   }
 
   @Test
-  void shouldBeAuthorizedToBroadcastSignalWithDefaultUser() {
+  void shouldBeAuthorizedToCompleteJobWithDefaultUser() {
     // given
-    createProcessInstance();
+    final var processInstanceKey =
+        defaultUserClient
+            .newCreateInstanceCommand()
+            .bpmnProcessId(PROCESS_ID)
+            .latestVersion()
+            .send()
+            .join()
+            .getProcessInstanceKey();
+    final var jobKey =
+        RecordingExporter.jobRecords(JobIntent.CREATED)
+            .withProcessInstanceKey(processInstanceKey)
+            .getFirst()
+            .getKey();
 
-    // when
-    final var response =
-        defaultUserClient.newBroadcastSignalCommand().signalName(SIGNAL_NAME).send().join();
-
-    // then
-    assertThat(response.getKey()).isPositive();
+    // when then
+    final var response = defaultUserClient.newCompleteCommand(jobKey).send().join();
+    // The Rest API returns a null future for an empty response
+    // We can verify for null, as if we'd be unauthenticated we'd get an exception
+    assertThat(response).isNull();
   }
 
   @Test
-  void shouldBeAuthorizedToBroadcastSignalWithUser() {
+  void shouldBeAuthorizedToCompleteJobWithUser() {
     // given
-    createProcessInstance();
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUserWithPermissions(
@@ -111,62 +115,53 @@ public class SignalBroadcastAuthorizationIT {
         new Permissions(
             ResourceTypeEnum.PROCESS_DEFINITION,
             PermissionTypeEnum.UPDATE_PROCESS_INSTANCE,
-            List.of(PROCESS_ID)),
-        new Permissions(
-            ResourceTypeEnum.PROCESS_DEFINITION,
-            PermissionTypeEnum.CREATE_PROCESS_INSTANCE,
             List.of(PROCESS_ID)));
+    final var processInstanceKey =
+        defaultUserClient
+            .newCreateInstanceCommand()
+            .bpmnProcessId(PROCESS_ID)
+            .latestVersion()
+            .send()
+            .join()
+            .getProcessInstanceKey();
+    final var jobKey =
+        RecordingExporter.jobRecords(JobIntent.CREATED)
+            .withProcessInstanceKey(processInstanceKey)
+            .getFirst()
+            .getKey();
 
     try (final var client = authUtil.createClient(username, password)) {
-      // when
-      final var response = client.newBroadcastSignalCommand().signalName(SIGNAL_NAME).send().join();
-
-      // then
-      assertThat(response.getKey()).isPositive();
+      // when then
+      final var response = client.newCompleteCommand(jobKey).send().join();
+      // The Rest API returns a null future for an empty response
+      // We can verify for null, as if we'd be unauthenticated we'd get an exception
+      assertThat(response).isNull();
     }
   }
 
   @Test
-  void shouldBeUnauthorizedToBroadcastSignalIfNoPermissions() {
+  void shouldBeUnauthorizedToCompleteJobIfNoPermissions() {
     // given
-    createProcessInstance();
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUser(username, password);
-
-    try (final var client = authUtil.createClient(username, password)) {
-
-      // when
-      final var response = client.newBroadcastSignalCommand().signalName(SIGNAL_NAME).send();
-
-      // then
-      assertThatThrownBy(response::join)
-          .isInstanceOf(ProblemException.class)
-          .hasMessageContaining("title: FORBIDDEN")
-          .hasMessageContaining("status: 403")
-          .hasMessageContaining(
-              "Insufficient permissions to perform operation 'CREATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
-              PROCESS_ID);
-    }
-  }
-
-  @Test
-  void shouldNotBroadcastSignalIfUnauthorizedForOne() {
-    // given
-    final var username = UUID.randomUUID().toString();
-    final var password = "password";
-    authUtil.createUserWithPermissions(
-        username,
-        password,
-        new Permissions(
-            ResourceTypeEnum.PROCESS_DEFINITION,
-            PermissionTypeEnum.CREATE_PROCESS_INSTANCE,
-            List.of(PROCESS_ID)));
-    final var processInstanceKey = createProcessInstance();
+    final var processInstanceKey =
+        defaultUserClient
+            .newCreateInstanceCommand()
+            .bpmnProcessId(PROCESS_ID)
+            .latestVersion()
+            .send()
+            .join()
+            .getProcessInstanceKey();
+    final var jobKey =
+        RecordingExporter.jobRecords(JobIntent.CREATED)
+            .withProcessInstanceKey(processInstanceKey)
+            .getFirst()
+            .getKey();
 
     try (final var client = authUtil.createClient(username, password)) {
       // when
-      final var response = client.newBroadcastSignalCommand().signalName(SIGNAL_NAME).send();
+      final var response = client.newCompleteCommand(jobKey).send();
 
       // then
       assertThatThrownBy(response::join)
@@ -176,36 +171,6 @@ public class SignalBroadcastAuthorizationIT {
           .hasMessageContaining(
               "Insufficient permissions to perform operation 'UPDATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
               PROCESS_ID);
-
-      assertThat(
-              RecordingExporter.records()
-                  .limit(r -> r.getRejectionType() == RejectionType.UNAUTHORIZED)
-                  .processInstanceRecords()
-                  .withProcessInstanceKey(processInstanceKey)
-                  .withElementType(BpmnElementType.INTERMEDIATE_CATCH_EVENT)
-                  .withEventType(BpmnEventType.SIGNAL)
-                  .withIntent(ProcessInstanceIntent.ELEMENT_COMPLETED)
-                  .exists())
-          .isFalse();
     }
-  }
-
-  private Long createProcessInstance() {
-    final var processInstanceKey =
-        defaultUserClient
-            .newCreateInstanceCommand()
-            .bpmnProcessId(PROCESS_ID)
-            .latestVersion()
-            .send()
-            .join()
-            .getProcessInstanceKey();
-
-    RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-        .withProcessInstanceKey(processInstanceKey)
-        .withElementType(BpmnElementType.INTERMEDIATE_CATCH_EVENT)
-        .withEventType(BpmnEventType.SIGNAL)
-        .await();
-
-    return processInstanceKey;
   }
 }
