+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/JobUpdateAuthorizationIT.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.it.authorization;
+package io.camunda.zeebe.engine.processing.authorization.permissions;
 
 import static io.camunda.zeebe.it.util.AuthorizationsUtil.createClient;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -18,11 +18,14 @@ import io.camunda.client.protocol.rest.PermissionTypeEnum;
 import io.camunda.client.protocol.rest.ResourceTypeEnum;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
+import io.camunda.zeebe.model.bpmn.Bpmn;
+import io.camunda.zeebe.protocol.record.intent.JobIntent;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
+import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
 import java.util.List;
 import java.util.UUID;
@@ -35,16 +38,19 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class RoleCreateAuthorizationIT {
+public class JobUpdateAuthorizationIT {
+
   @Container
   private static final ElasticsearchContainer CONTAINER =
       TestSearchContainers.createDefeaultElasticsearchContainer();
 
+  private static final String PROCESS_ID = "processId";
   private static AuthorizationsUtil authUtil;
   @AutoCloseResource private static CamundaClient defaultUserClient;
+  private static long jobKey;
 
   @TestZeebe(autoStart = false)
-  private final TestStandaloneBroker broker =
+  private TestStandaloneBroker broker =
       new TestStandaloneBroker()
           .withRecordingExporter(true)
           .withSecurityConfig(c -> c.getAuthorizations().setEnabled(true))
@@ -60,46 +66,75 @@ public class RoleCreateAuthorizationIT {
     authUtil = new AuthorizationsUtil(broker, defaultUserClient, CONTAINER.getHttpHostAddress());
 
     authUtil.awaitUserExistsInElasticsearch(defaultUsername);
+    defaultUserClient
+        .newDeployResourceCommand()
+        .addProcessModel(
+            Bpmn.createExecutableProcess(PROCESS_ID)
+                .startEvent()
+                .serviceTask("serviceTask", t -> t.zeebeJobType("jobType"))
+                .endEvent()
+                .done(),
+            "process.xml")
+        .send()
+        .join();
+
+    final var processInstanceKey =
+        defaultUserClient
+            .newCreateInstanceCommand()
+            .bpmnProcessId(PROCESS_ID)
+            .latestVersion()
+            .send()
+            .join()
+            .getProcessInstanceKey();
+    jobKey =
+        RecordingExporter.jobRecords(JobIntent.CREATED)
+            .withProcessInstanceKey(processInstanceKey)
+            .getFirst()
+            .getKey();
   }
 
   @Test
-  void shouldBeAuthorizedToCreateRoleWithDefaultUser() {
-    // when
-    final var response = defaultUserClient.newCreateRoleCommand().name("Foo").send().join();
-
-    // then
-    assertThat(response.getRoleKey()).isPositive();
+  void shouldBeAuthorizedToUpdateJobWithDefaultUser() {
+    // when then
+    final var response =
+        defaultUserClient.newUpdateJobCommand(jobKey).updateRetries(1).send().join();
+    // The Rest API returns a null future for an empty response
+    // We can verify for null, as if we'd be unauthenticated we'd get an exception
+    assertThat(response).isNull();
   }
 
   @Test
-  void shouldBeAuthorizedToCreateRoleWithPermissions() {
+  void shouldBeAuthorizedToUpdateJobWithUser() {
     // given
-    final var authUsername = UUID.randomUUID().toString();
+    final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUserWithPermissions(
-        authUsername,
+        username,
         password,
-        new Permissions(ResourceTypeEnum.ROLE, PermissionTypeEnum.CREATE, List.of("*")));
+        new Permissions(
+            ResourceTypeEnum.PROCESS_DEFINITION,
+            PermissionTypeEnum.UPDATE_PROCESS_INSTANCE,
+            List.of(PROCESS_ID)));
 
-    try (final var client = authUtil.createClient(authUsername, password)) {
-      // when
-      final var response = client.newCreateRoleCommand().name("Foo").send().join();
-
-      // then
-      assertThat(response.getRoleKey()).isPositive();
+    try (final var client = authUtil.createClient(username, password)) {
+      // when then
+      final var response = client.newUpdateJobCommand(jobKey).updateRetries(1).send().join();
+      // The Rest API returns a null future for an empty response
+      // We can verify for null, as if we'd be unauthenticated we'd get an exception
+      assertThat(response).isNull();
     }
   }
 
   @Test
-  void shouldBeUnAuthorizedToCreateRoleWithoutPermissions() {
+  void shouldBeUnauthorizedToUpdateJobIfNoPermissions() {
     // given
-    final var authUsername = UUID.randomUUID().toString();
+    final var username = UUID.randomUUID().toString();
     final var password = "password";
-    authUtil.createUser(authUsername, password);
+    authUtil.createUser(username, password);
 
-    // when
-    try (final var client = authUtil.createClient(authUsername, password)) {
-      final var response = client.newCreateRoleCommand().name("Foo").send();
+    try (final var client = authUtil.createClient(username, password)) {
+      // when
+      final var response = client.newUpdateJobCommand(jobKey).updateRetries(1).send();
 
       // then
       assertThatThrownBy(response::join)
@@ -107,7 +142,8 @@ public class RoleCreateAuthorizationIT {
           .hasMessageContaining("title: FORBIDDEN")
           .hasMessageContaining("status: 403")
           .hasMessageContaining(
-              "Insufficient permissions to perform operation 'CREATE' on resource 'ROLE'");
+              "Insufficient permissions to perform operation 'UPDATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              PROCESS_ID);
     }
   }
 }
