+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/SignalBroadcastAuthorizationIT.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.it.authorization;
+package io.camunda.zeebe.engine.processing.authorization.permissions;
 
 import static io.camunda.zeebe.it.util.AuthorizationsUtil.createClient;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -19,7 +19,10 @@ import io.camunda.client.protocol.rest.ResourceTypeEnum;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
 import io.camunda.zeebe.model.bpmn.Bpmn;
-import io.camunda.zeebe.protocol.record.intent.UserTaskIntent;
+import io.camunda.zeebe.protocol.record.RejectionType;
+import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
+import io.camunda.zeebe.protocol.record.value.BpmnElementType;
+import io.camunda.zeebe.protocol.record.value.BpmnEventType;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
@@ -38,8 +41,8 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class UserTaskClaimAuthorizationIT {
-  public static final String USER_TASK_ID = "userTask";
+public class SignalBroadcastAuthorizationIT {
+  public static final String SIGNAL_NAME = "signal";
 
   @Container
   private static final ElasticsearchContainer CONTAINER =
@@ -71,9 +74,12 @@ public class UserTaskClaimAuthorizationIT {
         .addProcessModel(
             Bpmn.createExecutableProcess(PROCESS_ID)
                 .startEvent()
-                .userTask(USER_TASK_ID)
-                .zeebeUserTask()
+                .intermediateCatchEvent()
+                .signal(s -> s.name(SIGNAL_NAME))
                 .endEvent()
+                .moveToProcess(PROCESS_ID)
+                .startEvent()
+                .signal(s -> s.name(SIGNAL_NAME))
                 .done(),
             "process.xml")
         .send()
@@ -81,25 +87,22 @@ public class UserTaskClaimAuthorizationIT {
   }
 
   @Test
-  void shouldBeAuthorizedToClaimUserTaskWithDefaultUser() {
+  void shouldBeAuthorizedToBroadcastSignalWithDefaultUser() {
     // given
-    final var processInstanceKey = createProcessInstance();
-    final var userTaskKey = getUserTaskKey(processInstanceKey);
+    createProcessInstance();
 
-    // when then
+    // when
     final var response =
-        defaultUserClient.newUserTaskAssignCommand(userTaskKey).assignee("assignee").send().join();
+        defaultUserClient.newBroadcastSignalCommand().signalName(SIGNAL_NAME).send().join();
 
-    // The Rest API returns a null future for an empty response
-    // We can verify for null, as if we'd be unauthenticated we'd get an exception
-    assertThat(response).isNull();
+    // then
+    assertThat(response.getKey()).isPositive();
   }
 
   @Test
-  void shouldBeAuthorizedToClaimUserTaskWithUser() {
+  void shouldBeAuthorizedToBroadcastSignalWithUser() {
     // given
-    final var processInstanceKey = createProcessInstance();
-    final var userTaskKey = getUserTaskKey(processInstanceKey);
+    createProcessInstance();
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUserWithPermissions(
@@ -107,32 +110,34 @@ public class UserTaskClaimAuthorizationIT {
         password,
         new Permissions(
             ResourceTypeEnum.PROCESS_DEFINITION,
-            PermissionTypeEnum.UPDATE_USER_TASK,
+            PermissionTypeEnum.UPDATE_PROCESS_INSTANCE,
+            List.of(PROCESS_ID)),
+        new Permissions(
+            ResourceTypeEnum.PROCESS_DEFINITION,
+            PermissionTypeEnum.CREATE_PROCESS_INSTANCE,
             List.of(PROCESS_ID)));
 
     try (final var client = authUtil.createClient(username, password)) {
-      // when then
-      final var response =
-          client.newUserTaskAssignCommand(userTaskKey).assignee("assignee").send().join();
+      // when
+      final var response = client.newBroadcastSignalCommand().signalName(SIGNAL_NAME).send().join();
 
-      // The Rest API returns a null future for an empty response
-      // We can verify for null, as if we'd be unauthenticated we'd get an exception
-      assertThat(response).isNull();
+      // then
+      assertThat(response.getKey()).isPositive();
     }
   }
 
   @Test
-  void shouldBeUnauthorizedToClaimUserTaskIfNoPermissions() {
+  void shouldBeUnauthorizedToBroadcastSignalIfNoPermissions() {
     // given
-    final var processInstanceKey = createProcessInstance();
-    final var userTaskKey = getUserTaskKey(processInstanceKey);
+    createProcessInstance();
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUser(username, password);
 
     try (final var client = authUtil.createClient(username, password)) {
-      // when we use the unauthorized client
-      final var response = client.newUserTaskAssignCommand(userTaskKey).assignee("assignee").send();
+
+      // when
+      final var response = client.newBroadcastSignalCommand().signalName(SIGNAL_NAME).send();
 
       // then
       assertThatThrownBy(response::join)
@@ -140,28 +145,67 @@ public class UserTaskClaimAuthorizationIT {
           .hasMessageContaining("title: FORBIDDEN")
           .hasMessageContaining("status: 403")
           .hasMessageContaining(
-              "Insufficient permissions to perform operation 'UPDATE_USER_TASK' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              "Insufficient permissions to perform operation 'CREATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
               PROCESS_ID);
     }
   }
 
-  private long createProcessInstance() {
-    return defaultUserClient
-        .newCreateInstanceCommand()
-        .bpmnProcessId(PROCESS_ID)
-        .latestVersion()
-        .send()
-        .join()
-        .getProcessInstanceKey();
+  @Test
+  void shouldNotBroadcastSignalIfUnauthorizedForOne() {
+    // given
+    final var username = UUID.randomUUID().toString();
+    final var password = "password";
+    authUtil.createUserWithPermissions(
+        username,
+        password,
+        new Permissions(
+            ResourceTypeEnum.PROCESS_DEFINITION,
+            PermissionTypeEnum.CREATE_PROCESS_INSTANCE,
+            List.of(PROCESS_ID)));
+    final var processInstanceKey = createProcessInstance();
+
+    try (final var client = authUtil.createClient(username, password)) {
+      // when
+      final var response = client.newBroadcastSignalCommand().signalName(SIGNAL_NAME).send();
+
+      // then
+      assertThatThrownBy(response::join)
+          .isInstanceOf(ProblemException.class)
+          .hasMessageContaining("title: FORBIDDEN")
+          .hasMessageContaining("status: 403")
+          .hasMessageContaining(
+              "Insufficient permissions to perform operation 'UPDATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              PROCESS_ID);
+
+      assertThat(
+              RecordingExporter.records()
+                  .limit(r -> r.getRejectionType() == RejectionType.UNAUTHORIZED)
+                  .processInstanceRecords()
+                  .withProcessInstanceKey(processInstanceKey)
+                  .withElementType(BpmnElementType.INTERMEDIATE_CATCH_EVENT)
+                  .withEventType(BpmnEventType.SIGNAL)
+                  .withIntent(ProcessInstanceIntent.ELEMENT_COMPLETED)
+                  .exists())
+          .isFalse();
+    }
   }
 
-  private static long getUserTaskKey(final long processInstanceKey) {
-    return RecordingExporter.userTaskRecords(UserTaskIntent.CREATED)
+  private Long createProcessInstance() {
+    final var processInstanceKey =
+        defaultUserClient
+            .newCreateInstanceCommand()
+            .bpmnProcessId(PROCESS_ID)
+            .latestVersion()
+            .send()
+            .join()
+            .getProcessInstanceKey();
+
+    RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
         .withProcessInstanceKey(processInstanceKey)
-        .withElementId(USER_TASK_ID)
-        .limit(1)
-        .findFirst()
-        .orElseThrow()
-        .getKey();
+        .withElementType(BpmnElementType.INTERMEDIATE_CATCH_EVENT)
+        .withEventType(BpmnEventType.SIGNAL)
+        .await();
+
+    return processInstanceKey;
   }
 }
