+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/UserTaskClaimAuthorizationIT.java
@@ -1,0 +1,167 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+package io.camunda.zeebe.engine.processing.authorization.permissions;
+
+import static io.camunda.zeebe.it.util.AuthorizationsUtil.createClient;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import io.camunda.application.Profile;
+import io.camunda.client.CamundaClient;
+import io.camunda.client.api.command.ProblemException;
+import io.camunda.client.protocol.rest.PermissionTypeEnum;
+import io.camunda.client.protocol.rest.ResourceTypeEnum;
+import io.camunda.zeebe.it.util.AuthorizationsUtil;
+import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
+import io.camunda.zeebe.model.bpmn.Bpmn;
+import io.camunda.zeebe.protocol.record.intent.UserTaskIntent;
+import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
+import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
+import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
+import io.camunda.zeebe.test.util.junit.AutoCloseResources;
+import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
+import io.camunda.zeebe.test.util.record.RecordingExporter;
+import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
+import java.util.List;
+import java.util.UUID;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.testcontainers.elasticsearch.ElasticsearchContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+@AutoCloseResources
+@Testcontainers
+@ZeebeIntegration
+public class UserTaskClaimAuthorizationIT {
+  public static final String USER_TASK_ID = "userTask";
+
+  @Container
+  private static final ElasticsearchContainer CONTAINER =
+      TestSearchContainers.createDefeaultElasticsearchContainer();
+
+  private static final String PROCESS_ID = "processId";
+  private static AuthorizationsUtil authUtil;
+  @AutoCloseResource private static CamundaClient defaultUserClient;
+
+  @TestZeebe(autoStart = false)
+  private TestStandaloneBroker broker =
+      new TestStandaloneBroker()
+          .withRecordingExporter(true)
+          .withSecurityConfig(c -> c.getAuthorizations().setEnabled(true))
+          .withAdditionalProfile(Profile.AUTH_BASIC);
+
+  @BeforeEach
+  void beforeEach() {
+    broker.withCamundaExporter("http://" + CONTAINER.getHttpHostAddress());
+    broker.start();
+
+    final var defaultUsername = "demo";
+    defaultUserClient = createClient(broker, defaultUsername, "demo");
+    authUtil = new AuthorizationsUtil(broker, defaultUserClient, CONTAINER.getHttpHostAddress());
+
+    authUtil.awaitUserExistsInElasticsearch(defaultUsername);
+    defaultUserClient
+        .newDeployResourceCommand()
+        .addProcessModel(
+            Bpmn.createExecutableProcess(PROCESS_ID)
+                .startEvent()
+                .userTask(USER_TASK_ID)
+                .zeebeUserTask()
+                .endEvent()
+                .done(),
+            "process.xml")
+        .send()
+        .join();
+  }
+
+  @Test
+  void shouldBeAuthorizedToClaimUserTaskWithDefaultUser() {
+    // given
+    final var processInstanceKey = createProcessInstance();
+    final var userTaskKey = getUserTaskKey(processInstanceKey);
+
+    // when then
+    final var response =
+        defaultUserClient.newUserTaskAssignCommand(userTaskKey).assignee("assignee").send().join();
+
+    // The Rest API returns a null future for an empty response
+    // We can verify for null, as if we'd be unauthenticated we'd get an exception
+    assertThat(response).isNull();
+  }
+
+  @Test
+  void shouldBeAuthorizedToClaimUserTaskWithUser() {
+    // given
+    final var processInstanceKey = createProcessInstance();
+    final var userTaskKey = getUserTaskKey(processInstanceKey);
+    final var username = UUID.randomUUID().toString();
+    final var password = "password";
+    authUtil.createUserWithPermissions(
+        username,
+        password,
+        new Permissions(
+            ResourceTypeEnum.PROCESS_DEFINITION,
+            PermissionTypeEnum.UPDATE_USER_TASK,
+            List.of(PROCESS_ID)));
+
+    try (final var client = authUtil.createClient(username, password)) {
+      // when then
+      final var response =
+          client.newUserTaskAssignCommand(userTaskKey).assignee("assignee").send().join();
+
+      // The Rest API returns a null future for an empty response
+      // We can verify for null, as if we'd be unauthenticated we'd get an exception
+      assertThat(response).isNull();
+    }
+  }
+
+  @Test
+  void shouldBeUnauthorizedToClaimUserTaskIfNoPermissions() {
+    // given
+    final var processInstanceKey = createProcessInstance();
+    final var userTaskKey = getUserTaskKey(processInstanceKey);
+    final var username = UUID.randomUUID().toString();
+    final var password = "password";
+    authUtil.createUser(username, password);
+
+    try (final var client = authUtil.createClient(username, password)) {
+      // when we use the unauthorized client
+      final var response = client.newUserTaskAssignCommand(userTaskKey).assignee("assignee").send();
+
+      // then
+      assertThatThrownBy(response::join)
+          .isInstanceOf(ProblemException.class)
+          .hasMessageContaining("title: FORBIDDEN")
+          .hasMessageContaining("status: 403")
+          .hasMessageContaining(
+              "Insufficient permissions to perform operation 'UPDATE_USER_TASK' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              PROCESS_ID);
+    }
+  }
+
+  private long createProcessInstance() {
+    return defaultUserClient
+        .newCreateInstanceCommand()
+        .bpmnProcessId(PROCESS_ID)
+        .latestVersion()
+        .send()
+        .join()
+        .getProcessInstanceKey();
+  }
+
+  private static long getUserTaskKey(final long processInstanceKey) {
+    return RecordingExporter.userTaskRecords(UserTaskIntent.CREATED)
+        .withProcessInstanceKey(processInstanceKey)
+        .withElementId(USER_TASK_ID)
+        .limit(1)
+        .findFirst()
+        .orElseThrow()
+        .getKey();
+  }
+}
