+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/UserTaskCompleteAuthorizationIT.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.it.authorization;
+package io.camunda.zeebe.engine.processing.authorization.permissions;
 
 import static io.camunda.zeebe.it.util.AuthorizationsUtil.createClient;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -19,14 +19,15 @@ import io.camunda.client.protocol.rest.ResourceTypeEnum;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
 import io.camunda.zeebe.model.bpmn.Bpmn;
+import io.camunda.zeebe.protocol.record.intent.UserTaskIntent;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
+import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
 import java.util.List;
-import java.util.Map;
 import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -37,10 +38,8 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class MessagePublishAuthorizationIT {
-  public static final String INTERMEDIATE_MSG_NAME = "intermediateMsg";
-  public static final String START_MSG_NAME = "startMsg";
-  public static final String CORRELATION_KEY_VARIABLE = "correlationKey";
+public class UserTaskCompleteAuthorizationIT {
+  public static final String USER_TASK_ID = "userTask";
 
   @Container
   private static final ElasticsearchContainer CONTAINER =
@@ -72,15 +71,9 @@ public class MessagePublishAuthorizationIT {
         .addProcessModel(
             Bpmn.createExecutableProcess(PROCESS_ID)
                 .startEvent()
-                .intermediateCatchEvent()
-                .message(
-                    m ->
-                        m.name(INTERMEDIATE_MSG_NAME)
-                            .zeebeCorrelationKeyExpression(CORRELATION_KEY_VARIABLE))
+                .userTask(USER_TASK_ID)
+                .zeebeUserTask()
                 .endEvent()
-                .moveToProcess(PROCESS_ID)
-                .startEvent()
-                .message(m -> m.name(START_MSG_NAME))
                 .done(),
             "process.xml")
         .send()
@@ -88,68 +81,56 @@ public class MessagePublishAuthorizationIT {
   }
 
   @Test
-  void shouldBeAuthorizedToPublishMessageWithDefaultUser() {
+  void shouldBeAuthorizedToCompleteUserTaskWithDefaultUser() {
     // given
-    final var correlationKey = UUID.randomUUID().toString();
-    createProcessInstance(correlationKey);
+    final var processInstanceKey = createProcessInstance();
+    final var userTaskKey = getUserTaskKey(processInstanceKey);
 
-    // when
-    final var response =
-        defaultUserClient
-            .newPublishMessageCommand()
-            .messageName(INTERMEDIATE_MSG_NAME)
-            .correlationKey(correlationKey)
-            .send()
-            .join();
+    // when then
+    final var response = defaultUserClient.newUserTaskCompleteCommand(userTaskKey).send().join();
 
-    // then
-    assertThat(response.getMessageKey()).isPositive();
+    // The Rest API returns a null future for an empty response
+    // We can verify for null, as if we'd be unauthenticated we'd get an exception
+    assertThat(response).isNull();
   }
 
   @Test
-  void shouldBeAuthorizedToPublishMessageWithUser() {
+  void shouldBeAuthorizedToCompleteUserTaskWithUser() {
     // given
-    final var correlationKey = UUID.randomUUID().toString();
-    createProcessInstance(correlationKey);
+    final var processInstanceKey = createProcessInstance();
+    final var userTaskKey = getUserTaskKey(processInstanceKey);
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUserWithPermissions(
         username,
         password,
-        new Permissions(ResourceTypeEnum.MESSAGE, PermissionTypeEnum.CREATE, List.of("*")));
+        new Permissions(
+            ResourceTypeEnum.PROCESS_DEFINITION,
+            PermissionTypeEnum.UPDATE_USER_TASK,
+            List.of(PROCESS_ID)));
 
     try (final var client = authUtil.createClient(username, password)) {
-      // when
-      final var response =
-          client
-              .newPublishMessageCommand()
-              .messageName(INTERMEDIATE_MSG_NAME)
-              .correlationKey(correlationKey)
-              .send()
-              .join();
+      // when then
+      final var response = client.newUserTaskCompleteCommand(userTaskKey).send().join();
 
-      // then
-      assertThat(response.getMessageKey()).isPositive();
+      // The Rest API returns a null future for an empty response
+      // We can verify for null, as if we'd be unauthenticated we'd get an exception
+      assertThat(response).isNull();
     }
   }
 
   @Test
-  void shouldBeUnauthorizedToPublishMessageIfNoPermissions() {
+  void shouldBeUnauthorizedToCompleteUserTaskIfNoPermissions() {
     // given
-    final var correlationKey = UUID.randomUUID().toString();
-    createProcessInstance(correlationKey);
+    final var processInstanceKey = createProcessInstance();
+    final var userTaskKey = getUserTaskKey(processInstanceKey);
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUser(username, password);
 
     try (final var client = authUtil.createClient(username, password)) {
-      // when
-      final var response =
-          client
-              .newPublishMessageCommand()
-              .messageName(INTERMEDIATE_MSG_NAME)
-              .correlationKey(correlationKey)
-              .send();
+      // when we use the unauthorized client
+      final var response = client.newUserTaskCompleteCommand(userTaskKey).send();
 
       // then
       assertThatThrownBy(response::join)
@@ -157,17 +138,28 @@ public class MessagePublishAuthorizationIT {
           .hasMessageContaining("title: FORBIDDEN")
           .hasMessageContaining("status: 403")
           .hasMessageContaining(
-              "Insufficient permissions to perform operation 'CREATE' on resource 'MESSAGE'");
+              "Insufficient permissions to perform operation 'UPDATE_USER_TASK' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              PROCESS_ID);
     }
   }
 
-  private void createProcessInstance(final String correlationKey) {
-    defaultUserClient
+  private long createProcessInstance() {
+    return defaultUserClient
         .newCreateInstanceCommand()
         .bpmnProcessId(PROCESS_ID)
         .latestVersion()
-        .variables(Map.of(CORRELATION_KEY_VARIABLE, correlationKey))
         .send()
-        .join();
+        .join()
+        .getProcessInstanceKey();
+  }
+
+  private static long getUserTaskKey(final long processInstanceKey) {
+    return RecordingExporter.userTaskRecords(UserTaskIntent.CREATED)
+        .withProcessInstanceKey(processInstanceKey)
+        .withElementId(USER_TASK_ID)
+        .limit(1)
+        .findFirst()
+        .orElseThrow()
+        .getKey();
   }
 }
